Fixed build for clang-14
We rely on code pointers staying at a fixed address.
When you think about it, there is probably no other (efficient)
way of doing it.

Since code pointers are values, they could end up in SKFS, and
therefore persist on disk. If another instance of the program
loading SKFS ends up with code pointers at a different address ...
well, things are going to go horribly wrong.

It turns out that fixed address for the code was the default for
clang until version 13, but now you have to specify -no-pie,
where pie stands for "position independent executable".

The other thing is that there is a *huge* regression for the
optimization time. I don't know why, it takes 4-5 times more time
to compile .ll files. The regression happened again starting
at clang-13.

So I am switching the default for optimization level from -O2 to
-O1. A lot of the optimizations performed by llvm are already done
in the SKIP backend anyway, so it should not affect performance all
that much.

Long term, we need to figure out what happened there ...